# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Lingow Account, Usage and Delivery API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
  /auth/anonymous:
    post:
      operationId: createAnonymousAccount
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResult' } } } }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /auth/verification-codes:
    post:
      operationId: createPhoneChallenge
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PhoneChallengeRequest' } } } }
      responses:
        '202': { description: Accepted, content: { application/json: { schema: { type: object, required: [challenge_id], properties: { challenge_id: { type: string } } } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /auth/phone/login:
    post:
      operationId: verifyPhone
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PhoneVerifyRequest' } } } }
      responses:
        '200': { description: Verified, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResult' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /auth/token/refresh:
    post:
      operationId: refreshToken
      requestBody: { $ref: '#/components/requestBodies/RefreshToken' }
      responses:
        '200': { description: Refreshed, content: { application/json: { schema: { $ref: '#/components/schemas/Tokens' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /auth/logout:
    post:
      operationId: logout
      requestBody: { $ref: '#/components/requestBodies/RefreshToken' }
      responses:
        '204': { description: Logged out }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /account/me:
    get:
      operationId: getMe
      security: [{ accountContext: [] }]
      responses:
        '200': { description: Current account, content: { application/json: { schema: { $ref: '#/components/schemas/Account' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { $ref: '#/components/responses/NotImplemented' }
...
  /voice-sessions/{id}/usage:
    get:
      operationId: getSessionUsage
      security: [{ accountContext: [] }]
      parameters: [{ $ref: '#/components/parameters/SessionID' }]
      responses:
        '200': { description: Session usage, content: { application/json: { schema: { $ref: '#/components/schemas/UsageSummary' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /usage/summary:
    get:
      operationId: getAccountUsage
      security: [{ accountContext: [] }]
      parameters:
        - { name: period_start, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: period_end, in: query, required: true, schema: { type: string, format: date-time } }
      responses:
        '200': { description: Account usage, content: { application/json: { schema: { $ref: '#/components/schemas/UsageSummary' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /outbound-messages:
    post:
      operationId: createOutboundMessage
      security: [{ accountContext: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateMessageRequest' } } } }
      responses:
        '202': { description: Queued, content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /outbound-messages/{message_id}:
    get:
      operationId: getOutboundMessage
      security: [{ accountContext: [] }]
      parameters: [{ $ref: '#/components/parameters/MessageID' }]
      responses:
        '200': { description: Message, content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { $ref: '#/components/responses/NotImplemented' }
...
  /account/message-preferences:
    get:
      operationId: listMessagePreferences
      security: [{ accountContext: [] }]
      responses:
        '200': { description: Preferences, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/Preference' } } } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /account/message-preferences/{channel}:
    put:
      operationId: putMessagePreference
      security: [{ accountContext: [] }]
      parameters: [{ name: channel, in: path, required: true, schema: { $ref: '#/components/schemas/Channel' } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, additionalProperties: false, required: [enabled], properties: { enabled: { type: boolean }, destination_ref: { type: string, description: 'One verified automatic destination for this channel' } } } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Preference' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /account/message-targets:
    get:
      operationId: listMessageTargets
      security: [{ accountContext: [] }]
      parameters: [{ name: channel, in: query, required: false, schema: { $ref: '#/components/schemas/Channel' } }]
      responses:
        '200': { description: Bound targets, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/MessageTarget' } } } } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /account/message-targets/wechat/bind:
    post:
      operationId: bindWeChatTarget
      security: [{ accountContext: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/BindWeChatRequest' } } } }
      responses:
        '200': { description: Bound target, content: { application/json: { schema: { $ref: '#/components/schemas/MessageTarget' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /account/message-targets/wechat/{destination_ref}:
    delete:
      operationId: unbindWeChatTarget
      security: [{ accountContext: [] }]
      parameters: [{ name: destination_ref, in: path, required: true, schema: { type: string } }]
      responses:
        '204': { description: Target revoked }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /account/message-targets/email/verification-codes:
    post:
      operationId: requestEmailBindVerification
      security: [{ accountContext: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email]
              properties:
                email: { type: string, minLength: 3, format: email }
                destination_ref: { type: string, minLength: 1 }
      responses:
        '202': { description: Verification token accepted for delivery }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /account/message-targets/email/bind:
    post:
      operationId: bindEmailTarget
      security: [{ accountContext: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/BindEmailRequest' } } } }
      responses:
        '200': { description: Bound target, content: { application/json: { schema: { $ref: '#/components/schemas/MessageTarget' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /account/message-targets/email/{destination_ref}:
    delete:
      operationId: unbindEmailTarget
      security: [{ accountContext: [] }]
      parameters: [{ name: destination_ref, in: path, required: true, schema: { type: string } }]
      responses:
        '204': { description: Target revoked }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '501': { $ref: '#/components/responses/NotImplemented' }
...
components:
  securitySchemes:
    accountContext:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token validated by authentication middleware; account_id comes from the trusted request context.
    realtimeTicket:
      type: http
      scheme: bearer
      description: HMAC-signed ticket authorizing access to one realtime Session.
  parameters:
    SessionID: { name: id, in: path, required: true, schema: { type: string } }
    MessageID: { name: message_id, in: path, required: true, schema: { type: string } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1, maxLength: 200 } }
  requestBodies:
    RefreshToken:
      required: true
      content: { application/json: { schema: { type: object, additionalProperties: false, required: [refresh_token], properties: { refresh_token: { type: string, minLength: 1 } } } } }
  responses:
    BadRequest: { description: Invalid request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unauthorized: { description: Missing or invalid authorization credentials, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Forbidden: { description: Authenticated account does not own the requested resource, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    TooManyRequests: { description: Phone verification requests are temporarily rate limited, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotImplemented: { description: Business implementation is intentionally pending, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  schemas:
...
    Account:
      type: object
      required: [id, kind, created_at]
      properties: { id: { type: string }, kind: { type: string, enum: [anonymous, registered] }, created_at: { type: string, format: date-time } }
    Tokens:
      type: object
      required: [access_token, refresh_token, expires_at]
      properties: { access_token: { type: string }, refresh_token: { type: string }, expires_at: { type: string, format: date-time } }
    AuthResult:
      type: object
      required: [account, tokens]
      properties: { account: { $ref: '#/components/schemas/Account' }, tokens: { $ref: '#/components/schemas/Tokens' } }
    PhoneChallengeRequest:
      type: object
      additionalProperties: false
      required: [phone]
      properties: { phone: { type: string, minLength: 1 } }
    PhoneVerifyRequest:
      type: object
      additionalProperties: false
      required: [challenge_id, code]
      properties:
        challenge_id: { type: string }
        code: { type: string }
        anonymous_account_id:
          type: string
          description: When supplied, the request must include a Bearer access token for this same anonymous account.
    Stage: { type: string, enum: [asr, translation, tts, diarization] }
    UsageSummary:
      type: object
      required: [account_id, period_start, period_end, totals]
      properties:
        account_id: { type: string }
        session_id: { type: string }
        period_start: { type: string, format: date-time }
        period_end: { type: string, format: date-time }
        totals:
          type: array
          items:
            type: object
            required: [service_type, input_tokens, output_tokens, audio_duration_ms, cost_amount, currency]
            properties:
              service_type: { $ref: '#/components/schemas/Stage' }
              input_tokens: { type: integer, format: int64, minimum: 0 }
              output_tokens: { type: integer, format: int64, minimum: 0 }
              audio_duration_ms: { type: integer, format: int64, minimum: 0 }
              cost_amount: { type: string }
              currency: { type: string }
    Channel: { type: string, enum: [email, wechat] }
    CreateMessageRequest:
      type: object
      additionalProperties: false
      required: [channel, destination_ref, turn_ids]
      properties: { channel: { $ref: '#/components/schemas/Channel' }, destination_ref: { type: string }, turn_ids: { type: array, minItems: 1, maxItems: 100, uniqueItems: true, items: { type: string } } }
    FinalTurnSnapshot:
      type: object
      required: [turn_id, session_id, participant_id, speaker_label_snapshot, source_language, target_language, language_config_version, source_text, translated_text, created_at]
      properties:
        turn_id: { type: string }
        session_id: { type: string }
        participant_id: { type: [string, 'null'] }
        speaker_label_snapshot: { type: [string, 'null'] }
        source_language: { type: string }
        target_language: { type: string }
        language_config_version: { type: integer, format: int64, minimum: 1 }
        source_text: { type: string }
        translated_text: { type: string }
        created_at: { type: string, format: date-time }
    Message:
      type: object
      required: [id, account_id, channel, destination_ref, snapshot_version, turns, status, attempts, last_error_code, created_at, updated_at]
      properties:
        id: { type: string }
        account_id: { type: string }
        channel: { $ref: '#/components/schemas/Channel' }
        destination_ref: { type: string }
        snapshot_version: { type: integer, minimum: 1 }
        turns: { type: array, items: { $ref: '#/components/schemas/FinalTurnSnapshot' } }
        status: { type: string, enum: [queued, sending, sent, failed, retrying, cancelled] }
        attempts: { type: integer, minimum: 0 }
        last_error_code: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Preference:
      type: object
      required: [account_id, channel, destination_ref, enabled, verified, updated_at]
      properties: { account_id: { type: string }, channel: { $ref: '#/components/schemas/Channel' }, destination_ref: { type: string }, enabled: { type: boolean }, verified: { type: boolean }, updated_at: { type: string, format: date-time } }
    MessageTarget:
      type: object
      required: [destination_ref, channel, verified, updated_at]
      properties:
        destination_ref: { type: string }
        channel: { $ref: '#/components/schemas/Channel' }
        verified: { type: boolean }
        revoked_at: { type: [string, 'null'], format: date-time }
        updated_at: { type: string, format: date-time }
    BindWeChatRequest:
      type: object
      additionalProperties: false
      required: [code]
      properties:
        code: { type: string, minLength: 1, description: 'WeChat OAuth login code exchanged server-side for OpenID' }
    BindEmailRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token: { type: string, minLength: 1, description: 'One-time email verification token; dev mode accepts dev: prefixed addresses' }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id, retryable, details]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }
            retryable: { type: boolean }
            details: { type: object, additionalProperties: true }